Fix triangle gap issue with clipping
Added a 2d clipper on the bounding box for triangle pixel testing (clipped to framebuffer size)

Removed X and Y 3d clippping (keeping Z clipping for verts behind the camera)

The main issue was floating point impresion when lerping with the old clippping

+++ src/Math/Point2.hpp
@@ -11,12 +11,12 @@ namespace Math
 	class Point2
 	{
 	public:
-		uint16_t x;
-		uint16_t y;
+		int32_t x;
+		int32_t y;
 
-		constexpr Point2(uint16_t x, uint16_t y) : x(x), y(y) {}
-		constexpr Point2(const Vector2& vector) : x((uint16_t)vector.x), y((uint16_t)vector.y) {}
-		constexpr Point2(const Vector3& vector) : x((uint16_t)vector.x), y((uint16_t)vector.y) {}
+		constexpr Point2(int32_t x, int32_t y) : x(x), y(y) {}
+		constexpr Point2(const Vector2& vector) : x((int32_t)vector.x), y((int32_t)vector.y) {}
+		constexpr Point2(const Vector3& vector) : x((int32_t)vector.x), y((int32_t)vector.y) {}
 
 		constexpr Point2& operator+=(const Point2& other) noexcept
 		{
